Add /map/change route; extract shared MapPage component
- /map → value bubble map (initialViewMode="value")
- /map/change → % change map (initialViewMode="change")
- +page.server.js promoted to +layout.server.js so both routes share data
- Toggle uses replaceState() to update URL without remounting / losing state
