+++ packages/web/src/routes/map/change/+page.svelte
@@ -1,0 +1,6 @@
+<script>
+  import MapPage from "../MapPage.svelte";
+  export let data;
+</script>
+
+<MapPage {data} initialViewMode="change" />
